3654 - move history to NavAssessment
