+++ src/client/components/Navigation/NavAssessment/NavAssessment.tsx
@@ -9,11 +9,13 @@ import { Objects } from 'utils/objects'
 import { Areas, CountryIso } from 'meta/area'
 import { Routes } from 'meta/routes'
 
+import { useIsHistoryActive } from 'client/store/data'
 import { useSections } from 'client/store/metadata'
 import { useCountryRouteParams } from 'client/hooks/useRouteParams'
 import Hr from 'client/components/Hr'
 import Icon from 'client/components/Icon'
 import Header from 'client/components/Navigation/NavAssessment/Header'
+import History from 'client/components/Navigation/NavAssessment/History'
 import NavigationSection from 'client/components/Navigation/NavAssessment/Section'
 import { Breakpoints } from 'client/utils'
 
@@ -23,6 +25,7 @@ const NavAssessment: React.FC = () => {
   const { t } = useTranslation()
   const { assessmentName, cycleName, countryIso } = useCountryRouteParams<CountryIso>()
   const sections = useSections()
+  const historyActive = useIsHistoryActive()
 
   const maxHeight = useMaxHeight()
   const [showSections, setShowSections] = useState<boolean>(false)
@@ -30,27 +33,32 @@ const NavAssessment: React.FC = () => {
 
   return (
     <div className="nav-assessment" style={{ maxHeight }}>
-      <Header setShowSections={setShowSections} showSections={showSections} />
+      {historyActive && <History />}
+      {!historyActive && (
+        <>
+          <Header setShowSections={setShowSections} showSections={showSections} />
 
-      {sections.map((section) => (
-        <NavigationSection key={section.uuid} section={section} showSections={showSections} />
-      ))}
+          {sections.map((section) => (
+            <NavigationSection key={section.uuid} section={section} showSections={showSections} />
+          ))}
 
-      {Areas.isGlobal(countryIso) && (
-        <MediaQuery minWidth={Breakpoints.laptop}>
-          <div className="nav-header__sep-container">
-            <Hr />
-          </div>
-          <Link
-            className="nav-section__header nav-assessment__bulk-download"
-            to={Routes.CountryDataDownload.generatePath({ assessmentName, cycleName, countryIso })}
-          >
-            <div className="nav-section__order">
-              <Icon className="icon-sub icon-white" name="hit-down" />
-            </div>
-            <div className="nav-section__label">{t('dataDownload.dataDownload')}</div>
-          </Link>
-        </MediaQuery>
+          {Areas.isGlobal(countryIso) && (
+            <MediaQuery minWidth={Breakpoints.laptop}>
+              <div className="nav-header__sep-container">
+                <Hr />
+              </div>
+              <Link
+                className="nav-section__header nav-assessment__bulk-download"
+                to={Routes.CountryDataDownload.generatePath({ assessmentName, cycleName, countryIso })}
+              >
+                <div className="nav-section__order">
+                  <Icon className="icon-sub icon-white" name="hit-down" />
+                </div>
+                <div className="nav-section__label">{t('dataDownload.dataDownload')}</div>
+              </Link>
+            </MediaQuery>
+          )}
+        </>
       )}
     </div>
   )
