+++ src/client/components/Navigation/Navigation.tsx
@@ -3,11 +3,9 @@ import React, { useEffect } from 'react'
 import MediaQuery from 'react-responsive'
 
 import { useAppDispatch } from 'client/store'
-import { useIsHistoryActive } from 'client/store/data'
 import { NavigationActions } from 'client/store/ui/navigation'
 import { useCountryIso } from 'client/hooks'
 import { useIsPrintRoute } from 'client/hooks/useIsRoute'
-import History from 'client/components/Navigation/History'
 import { Breakpoints } from 'client/utils'
 
 import NavigationDesktop from './NavigationDesktop'
@@ -17,14 +15,12 @@ const Navigation: React.FC = () => {
   const countryIso = useCountryIso()
   const dispatch = useAppDispatch()
   const { print } = useIsPrintRoute()
-  const historyActive = useIsHistoryActive()
 
   useEffect(() => {
     if (print) dispatch(NavigationActions.updateNavigationVisible(false))
   }, [print, dispatch])
 
   if (!countryIso || print) return <div />
-  if (historyActive) return <History />
 
   return (
     <>
